pyproject.toml
[tool.poetry]
name = "comics-finder-api"
version = "0.1.0"
description = "Comics Finder REST API build with the Django REST Framework"
authors = ["Florian DEHE <florian.dehe@utt.fr>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
